requirements.txt
# Deep Learning Based Low-Light Image Enhancement for Improved Visibility
# B.Tech Final Year Academic Project - Author: Vivek Katta
# Jyothismathi Institute of Technology and Science
# Requirements for CPU-compatible training and inference

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.19.0

# Web Framework
Flask>=2.0.0
Werkzeug>=2.0.0

# Data Processing and Visualization
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.62.0

# Additional Utilities
pathlib2>=2.3.0
